GhibliLens/backend/pyproject.toml
[project]
name = "ghibilens-backend"
version = "0.1.0"
description = "GhibliLens - AI-powered Studio Ghibli video stylizer backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "replicate>=0.30.0",
    "opencv-python>=4.9.0.80",
    "Pillow>=10.3.0",
    "ffmpeg-python>=0.2.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.3.0",
    "aiofiles>=23.2.1",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools]
packages = []

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "httpx>=0.27.0",
]
